+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - refactor resource listing
+- `package.json` `main` attribute outputs scan classes
 
 ## [1.6.1] - 2020-10-27